pyproject.toml
[tool.poetry]
name = "bingkit-ffxiv"
version = "0.3.0.pre0"
description = "개인용 FF14 툴킷"
authors = ["Bingsu <ks2515@naver.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pandas = { version = "^2.0", extras = ["excel"] }
polars = { version = "^0.19", extras = ["pandas"] }
tqdm = "*"
loguru = "*"
winloop = { version = "*", markers = "platform_system == 'Windows'" }
taskgroup = { version = "*", python = "<3.11" }

[tool.poetry.group.dev.dependencies]
black = "*"
ruff = "*"
pre-commit = "*"
ipykernel = "*"

[[tool.poetry.packages]]
include = "bingkit"
from = "src"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.ruff]
select = ["A", "B", "C4", "E", "F", "I001", "N", "PT", "UP", "W"]
ignore = ["B008", "B905", "E501"]
unfixable = ["F401"]
